backend/requirements.txt
# Django
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
django-filter>=23.0
psycopg2-binary>=2.9

# ML service
pydantic-settings>=2.0
numpy>=1.24
torch>=2.0
torch-geometric>=2.4
sentence-transformers>=2.2
python-dotenv>=1.0
python-jobspy>=1.1
pdfplumber>=0.10
python-docx>=1.0
datasets>=4.0

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.4
